requirement.txt
streamlit==1.25.0
langdetect==1.0.9
langid==1.1.6
googletrans==4.0.0rc1
PyPDF2==3.0.1
python-docx==0.8.11
gTTS==2.3.1
